Курсы,лекции,тесты

+++ src/main/java/ru/top/project/model/HomeWork.java
@@ -78,7 +78,7 @@ public class HomeWork {
     public BigInteger getId() {
         return id;
     }
-    public void setId(int id) {
-        this.id = BigInteger.valueOf(id);
+    public void setId(BigInteger id) {
+        this.id = id;
     }
 }
